#43: Move loader files to bootloader files
